refactor(trading): consolidate market-signal orchestration behind facade (#2425)
Move the layer-① market-signal orchestration out of app-level glue in
crates/app/src/finance_event.rs into a cohesive facade in rara-trading
(crates/extensions/rara-trading/src/dispatch/). Behavior is bit-for-bit
preserved; this is a boundary-convergence refactor, not a feature.

The new `dispatch::on_feed_event` entry point runs the whole
persist -> upsert -> evaluate -> match -> deliver pipeline for one kernel
FeedEvent, returning FeedDispatchOutcome. It owns what was app-private
glue: the rolling-window pull + degradation policy (evaluate_anomaly),
candle parsing/normalization (market_candle_from_event), directive
wording (finance_directive), and the FeedDispatchSink trait +
FeedDispatchOutcome (both depend only on rara-kernel types). The parity
test suite moved with the code and now drives the facade entry point.

crates/app degrades to thin wiring: KernelFeedDispatchSink (wraps
KernelHandle, must not leak into rara-trading) stays in finance_event.rs,
and the feed-dispatch loop in lib.rs calls the facade entry point instead
of local helpers. The registry stays the single home of the
cooldown/budget/bypass delivery policy; the facade only supplies the
already-computed severity.

Closes #2425
