+++ crates/extensions/rara-trading/src/lib.rs
@@ -20,6 +20,7 @@
 //! [`FeedEvent`](rara_kernel::data_feed::FeedEvent) values.
 
 pub mod anomaly;
+pub mod dispatch;
 pub mod feed;
 pub mod finance;
 pub mod market_data;
